+++ src/main/java/vy/app/repository/MemberRepository.java
@@ -8,7 +8,7 @@ import vy.app.model.Member;
 
 import java.util.List;
 
-public interface MemberRepository extends PagingAndSortingRepository<Member, Integer>, JpaSpecificationExecutor<Member> {
+public interface MemberRepository extends PagingAndSortingRepository<Member, Long>, JpaSpecificationExecutor<Member> {
     //SELECT * from vy_member as m where m.last_name like '%singh%' or m.first_name LIKE '%singh%' or m.middle_name LIKE '%singh%'
 //    @Query("select m from Member m where m.firstName like concat('%', lower(?1), '%') or m.middleName like concat('%', lower(?1), '%') or m.lastName like concat('%', lower(?1), '%')")
 //    List<Member> getByName(String name);
